v1.4: World expansion update complete
FINALLY FINISHED THE DAMN WORLD EXPANSION UPDATE

Updates:
 - Gates can now be any size
 - Player enters a new area based on his position in previous area
 - Camera repositioning now works when going from world to world (and with teleportation)
 - Camera repositioning code is now looped instead of linear (meaning any teleportation shouldn't throw off camera)
 - other: face sprites need to be named with number prefix since they are retrieved alphabetically

TODO:
 - Inventory update

+++ MSJGame/core/src/woohoo/gameobjects/Player.java
@@ -66,7 +66,7 @@ public class Player extends Character
 		collision.setForce(0, 0);
 	}
     
-    public void setPosition(int x, int y)
+    public void setPosition(float x, float y)
     {
         collision.setPosition(x, y);
     }
